@@ -8,6 +8,7 @@ import android.app.Activity;
 import android.hardware.camera2.CameraAccessException;
 import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
+import io.flutter.embedding.engine.systemchannels.PlatformChannel;
 import io.flutter.plugin.common.BinaryMessenger;
 import io.flutter.plugin.common.EventChannel;
 import io.flutter.plugin.common.MethodCall;
@@ -255,7 +256,7 @@ final class MethodCallHandlerImpl implements MethodChannel.MethodCallHandler {
       case "stopImageStream":
         {
           try {
-            camera.stopImageStream();
+            camera.startPreview();
             result.success(null);
           } catch (Exception e) {
             handleException(e, result);
@@ -300,6 +301,29 @@ final class MethodCallHandlerImpl implements MethodChannel.MethodCallHandler {
 
           try {
             camera.setZoomLevel(result, zoom.floatValue());
+          } catch (Exception e) {
+            handleException(e, result);
+          }
+          break;
+        }
+      case "lockCaptureOrientation":
+        {
+          PlatformChannel.DeviceOrientation orientation =
+              CameraUtils.deserializeDeviceOrientation(call.argument("orientation"));
+
+          try {
+            camera.lockCaptureOrientation(orientation);
+            result.success(null);
+          } catch (Exception e) {
+            handleException(e, result);
+          }
+          break;
+        }
+      case "unlockCaptureOrientation":
+        {
+          try {
+            camera.unlockCaptureOrientation();
+            result.success(null);
           } catch (Exception e) {
             handleException(e, result);
           }
