@@ -38,6 +38,9 @@ get mockOnCameraInitializedEvent => CameraInitializedEvent(
       true,
     );
 
+get mockOnDeviceOrientationChangedEvent =>
+    DeviceOrientationChangedEvent(DeviceOrientation.portraitUp);
+
 get mockOnCameraClosingEvent => null;
 
 get mockOnCameraErrorEvent => CameraErrorEvent(13, 'closing');
@@ -210,12 +213,21 @@ void main() {
               sensorOrientation: 90),
           ResolutionPreset.max);
       expect(
-          cameraController.takePicture(),
-          throwsA(isA<CameraException>().having(
-            (error) => error.description,
-            'Uninitialized CameraController.',
-            'takePicture was called on uninitialized CameraController',
-          )));
+        cameraController.takePicture(),
+        throwsA(
+          isA<CameraException>()
+              .having(
+                (error) => error.code,
+                'code',
+                'Uninitialized CameraController',
+              )
+              .having(
+                (error) => error.description,
+                'description',
+                'takePicture() was called on an uninitialized CameraController.',
+              ),
+        ),
+      );
     });
 
     test('takePicture() throws $CameraException when takePicture is true',
@@ -283,12 +295,21 @@ void main() {
           ResolutionPreset.max);
 
       expect(
-          cameraController.startVideoRecording(),
-          throwsA(isA<CameraException>().having(
-            (error) => error.description,
-            'Uninitialized CameraController',
-            'startVideoRecording was called on uninitialized CameraController',
-          )));
+        cameraController.startVideoRecording(),
+        throwsA(
+          isA<CameraException>()
+              .having(
+                (error) => error.code,
+                'code',
+                'Uninitialized CameraController',
+              )
+              .having(
+                (error) => error.description,
+                'description',
+                'startVideoRecording() was called on an uninitialized CameraController.',
+              ),
+        ),
+      );
     });
     test('startVideoRecording() throws $CameraException when recording videos',
         () async {
@@ -347,12 +368,21 @@ void main() {
           ResolutionPreset.max);
 
       expect(
-          cameraController.getMaxZoomLevel,
-          throwsA(isA<CameraException>().having(
-            (error) => error.description,
-            'Uninitialized CameraController',
-            'getMaxZoomLevel was called on uninitialized CameraController',
-          )));
+        cameraController.getMaxZoomLevel,
+        throwsA(
+          isA<CameraException>()
+              .having(
+                (error) => error.code,
+                'code',
+                'Uninitialized CameraController',
+              )
+              .having(
+                (error) => error.description,
+                'description',
+                'getMaxZoomLevel() was called on an uninitialized CameraController.',
+              ),
+        ),
+      );
     });
 
     test('getMaxZoomLevel() throws $CameraException when disposed', () async {
@@ -367,12 +397,21 @@ void main() {
       await cameraController.dispose();
 
       expect(
-          cameraController.getMaxZoomLevel,
-          throwsA(isA<CameraException>().having(
-            (error) => error.description,
-            'Uninitialized CameraController',
-            'getMaxZoomLevel was called on uninitialized CameraController',
-          )));
+        cameraController.getMaxZoomLevel,
+        throwsA(
+          isA<CameraException>()
+              .having(
+                (error) => error.code,
+                'code',
+                'Disposed CameraController',
+              )
+              .having(
+                (error) => error.description,
+                'description',
+                'getMaxZoomLevel() was called on a disposed CameraController.',
+              ),
+        ),
+      );
     });
 
     test(
@@ -429,12 +468,21 @@ void main() {
           ResolutionPreset.max);
 
       expect(
-          cameraController.getMinZoomLevel,
-          throwsA(isA<CameraException>().having(
-            (error) => error.description,
-            'Uninitialized CameraController',
-            'getMinZoomLevel was called on uninitialized CameraController',
-          )));
+        cameraController.getMinZoomLevel,
+        throwsA(
+          isA<CameraException>()
+              .having(
+                (error) => error.code,
+                'code',
+                'Uninitialized CameraController',
+              )
+              .having(
+                (error) => error.description,
+                'description',
+                'getMinZoomLevel() was called on an uninitialized CameraController.',
+              ),
+        ),
+      );
     });
 
     test('getMinZoomLevel() throws $CameraException when disposed', () async {
@@ -449,12 +497,21 @@ void main() {
       await cameraController.dispose();
 
       expect(
-          cameraController.getMinZoomLevel,
-          throwsA(isA<CameraException>().having(
-            (error) => error.description,
-            'Uninitialized CameraController',
-            'getMinZoomLevel was called on uninitialized CameraController',
-          )));
+        cameraController.getMinZoomLevel,
+        throwsA(
+          isA<CameraException>()
+              .having(
+                (error) => error.code,
+                'code',
+                'Disposed CameraController',
+              )
+              .having(
+                (error) => error.description,
+                'description',
+                'getMinZoomLevel() was called on a disposed CameraController.',
+              ),
+        ),
+      );
     });
 
     test(
@@ -510,12 +567,21 @@ void main() {
           ResolutionPreset.max);
 
       expect(
-          () => cameraController.setZoomLevel(42.0),
-          throwsA(isA<CameraException>().having(
-            (error) => error.description,
-            'Uninitialized CameraController',
-            'setZoomLevel was called on uninitialized CameraController',
-          )));
+        () => cameraController.setZoomLevel(42.0),
+        throwsA(
+          isA<CameraException>()
+              .having(
+                (error) => error.code,
+                'code',
+                'Uninitialized CameraController',
+              )
+              .having(
+                (error) => error.description,
+                'description',
+                'setZoomLevel() was called on an uninitialized CameraController.',
+              ),
+        ),
+      );
     });
 
     test('setZoomLevel() throws $CameraException when disposed', () async {
@@ -530,12 +596,21 @@ void main() {
       await cameraController.dispose();
 
       expect(
-          () => cameraController.setZoomLevel(42.0),
-          throwsA(isA<CameraException>().having(
-            (error) => error.description,
-            'Uninitialized CameraController',
-            'setZoomLevel was called on uninitialized CameraController',
-          )));
+        () => cameraController.setZoomLevel(42.0),
+        throwsA(
+          isA<CameraException>()
+              .having(
+                (error) => error.code,
+                'code',
+                'Disposed CameraController',
+              )
+              .having(
+                (error) => error.description,
+                'description',
+                'setZoomLevel() was called on a disposed CameraController.',
+              ),
+        ),
+      );
     });
 
     test(
@@ -1021,6 +1096,106 @@ void main() {
               .setExposureOffset(cameraController.cameraId, -0.4))
           .called(4);
     });
+
+    test('lockCaptureOrientation() calls $CameraPlatform', () async {
+      CameraController cameraController = CameraController(
+          CameraDescription(
+              name: 'cam',
+              lensDirection: CameraLensDirection.back,
+              sensorOrientation: 90),
+          ResolutionPreset.max);
+      await cameraController.initialize();
+
+      await cameraController.lockCaptureOrientation();
+      expect(cameraController.value.lockedCaptureOrientation,
+          equals(DeviceOrientation.portraitUp));
+      await cameraController
+          .lockCaptureOrientation(DeviceOrientation.landscapeRight);
+      expect(cameraController.value.lockedCaptureOrientation,
+          equals(DeviceOrientation.landscapeRight));
+
+      verify(CameraPlatform.instance.lockCaptureOrientation(
+              cameraController.cameraId, DeviceOrientation.portraitUp))
+          .called(1);
+      verify(CameraPlatform.instance.lockCaptureOrientation(
+              cameraController.cameraId, DeviceOrientation.landscapeRight))
+          .called(1);
+    });
+
+    test(
+        'lockCaptureOrientation() throws $CameraException on $PlatformException',
+        () async {
+      CameraController cameraController = CameraController(
+          CameraDescription(
+              name: 'cam',
+              lensDirection: CameraLensDirection.back,
+              sensorOrientation: 90),
+          ResolutionPreset.max);
+      await cameraController.initialize();
+      when(CameraPlatform.instance.lockCaptureOrientation(
+              cameraController.cameraId, DeviceOrientation.portraitUp))
+          .thenThrow(
+        PlatformException(
+          code: 'TEST_ERROR',
+          message: 'This is a test error message',
+          details: null,
+        ),
+      );
+
+      expect(
+          cameraController.lockCaptureOrientation(DeviceOrientation.portraitUp),
+          throwsA(isA<CameraException>().having(
+            (error) => error.description,
+            'TEST_ERROR',
+            'This is a test error message',
+          )));
+    });
+
+    test('unlockCaptureOrientation() calls $CameraPlatform', () async {
+      CameraController cameraController = CameraController(
+          CameraDescription(
+              name: 'cam',
+              lensDirection: CameraLensDirection.back,
+              sensorOrientation: 90),
+          ResolutionPreset.max);
+      await cameraController.initialize();
+
+      await cameraController.unlockCaptureOrientation();
+      expect(cameraController.value.lockedCaptureOrientation, equals(null));
+
+      verify(CameraPlatform.instance
+              .unlockCaptureOrientation(cameraController.cameraId))
+          .called(1);
+    });
+
+    test(
+        'unlockCaptureOrientation() throws $CameraException on $PlatformException',
+        () async {
+      CameraController cameraController = CameraController(
+          CameraDescription(
+              name: 'cam',
+              lensDirection: CameraLensDirection.back,
+              sensorOrientation: 90),
+          ResolutionPreset.max);
+      await cameraController.initialize();
+      when(CameraPlatform.instance
+              .unlockCaptureOrientation(cameraController.cameraId))
+          .thenThrow(
+        PlatformException(
+          code: 'TEST_ERROR',
+          message: 'This is a test error message',
+          details: null,
+        ),
+      );
+
+      expect(
+          cameraController.unlockCaptureOrientation(),
+          throwsA(isA<CameraException>().having(
+            (error) => error.description,
+            'TEST_ERROR',
+            'This is a test error message',
+          )));
+    });
   });
 }
 
@@ -1056,6 +1231,10 @@ class MockCameraPlatform extends Mock
   @override
   Stream<CameraErrorEvent> onCameraError(int cameraId) =>
       Stream.value(mockOnCameraErrorEvent);
+
+  @override
+  Stream<DeviceOrientationChangedEvent> onDeviceOrientationChanged() =>
+      Stream.value(mockOnDeviceOrientationChangedEvent);
 
   @override
   Future<XFile> takePicture(int cameraId) => mockPlatformException
