@@ -7,22 +7,27 @@ import 'dart:ui';
 import 'package:camera/camera.dart';
 import 'package:camera_platform_interface/camera_platform_interface.dart';
 import 'package:flutter/cupertino.dart';
+import 'package:flutter/services.dart';
 import 'package:flutter_test/flutter_test.dart';
 
 void main() {
   group('camera_value', () {
     test('Can be created', () {
       var cameraValue = const CameraValue(
-          isInitialized: false,
-          errorDescription: null,
-          previewSize: Size(10, 10),
-          isRecordingPaused: false,
-          isRecordingVideo: false,
-          isTakingPicture: false,
-          isStreamingImages: false,
-          flashMode: FlashMode.auto,
-          exposureMode: ExposureMode.auto,
-          exposurePointSupported: true);
+        isInitialized: false,
+        errorDescription: null,
+        previewSize: Size(10, 10),
+        isRecordingPaused: false,
+        isRecordingVideo: false,
+        isTakingPicture: false,
+        isStreamingImages: false,
+        flashMode: FlashMode.auto,
+        exposureMode: ExposureMode.auto,
+        exposurePointSupported: true,
+        deviceOrientation: DeviceOrientation.portraitUp,
+        lockedCaptureOrientation: DeviceOrientation.portraitUp,
+        recordingOrientation: DeviceOrientation.portraitUp,
+      );
 
       expect(cameraValue, isA<CameraValue>());
       expect(cameraValue.isInitialized, isFalse);
@@ -35,6 +40,10 @@ void main() {
       expect(cameraValue.flashMode, FlashMode.auto);
       expect(cameraValue.exposureMode, ExposureMode.auto);
       expect(cameraValue.exposurePointSupported, true);
+      expect(cameraValue.deviceOrientation, DeviceOrientation.portraitUp);
+      expect(
+          cameraValue.lockedCaptureOrientation, DeviceOrientation.portraitUp);
+      expect(cameraValue.recordingOrientation, DeviceOrientation.portraitUp);
     });
 
     test('Can be created as uninitialized', () {
@@ -51,6 +60,9 @@ void main() {
       expect(cameraValue.flashMode, FlashMode.auto);
       expect(cameraValue.exposureMode, null);
       expect(cameraValue.exposurePointSupported, false);
+      expect(cameraValue.deviceOrientation, DeviceOrientation.portraitUp);
+      expect(cameraValue.lockedCaptureOrientation, null);
+      expect(cameraValue.recordingOrientation, null);
     });
 
     test('Can be copied with isInitialized', () {
@@ -68,6 +80,9 @@ void main() {
       expect(cameraValue.flashMode, FlashMode.auto);
       expect(cameraValue.exposureMode, null);
       expect(cameraValue.exposurePointSupported, false);
+      expect(cameraValue.deviceOrientation, DeviceOrientation.portraitUp);
+      expect(cameraValue.lockedCaptureOrientation, null);
+      expect(cameraValue.recordingOrientation, null);
     });
 
     test('Has aspectRatio after setting size', () {
@@ -75,7 +90,7 @@ void main() {
       var cameraValue =
           cv.copyWith(isInitialized: true, previewSize: Size(20, 10));
 
-      expect(cameraValue.aspectRatio, 0.5);
+      expect(cameraValue.aspectRatio, 2.0);
     });
 
     test('hasError is true after setting errorDescription', () {
@@ -110,10 +125,13 @@ void main() {
         focusMode: FocusMode.auto,
         exposurePointSupported: true,
         focusPointSupported: true,
+        deviceOrientation: DeviceOrientation.portraitUp,
+        lockedCaptureOrientation: DeviceOrientation.portraitUp,
+        recordingOrientation: DeviceOrientation.portraitUp,
       );
 
       expect(cameraValue.toString(),
-          'CameraValue(isRecordingVideo: false, isInitialized: false, errorDescription: null, previewSize: Size(10.0, 10.0), isStreamingImages: false, flashMode: FlashMode.auto, exposureMode: ExposureMode.auto, focusMode: FocusMode.auto, exposurePointSupported: true, focusPointSupported: true)');
+          'CameraValue(isRecordingVideo: false, isInitialized: false, errorDescription: null, previewSize: Size(10.0, 10.0), isStreamingImages: false, flashMode: FlashMode.auto, exposureMode: ExposureMode.auto, focusMode: FocusMode.auto, exposurePointSupported: true, focusPointSupported: true, deviceOrientation: DeviceOrientation.portraitUp, lockedCaptureOrientation: DeviceOrientation.portraitUp, recordingOrientation: DeviceOrientation.portraitUp)');
     });
   });
 }
